feat(infra_alert_condition): add violation_close_timer to newrelic_infra_alert_condition resource
