+++ newrelic/resource_newrelic_infra_alert_condition.go
@@ -1,12 +1,10 @@
 package newrelic
 
 import (
-	"fmt"
 	"log"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
-	"github.com/newrelic/newrelic-client-go/pkg/alerts"
 	"github.com/newrelic/newrelic-client-go/pkg/errors"
 )
 
@@ -136,138 +134,14 @@ func resourceNewRelicInfraAlertCondition() *schema.Resource {
 				Type:     schema.TypeString,
 				Optional: true,
 			},
+			"violation_close_timer": {
+				Type:         schema.TypeInt,
+				Optional:     true,
+				Default:      24,
+				ValidateFunc: intInSlice([]int{0, 1, 2, 4, 8, 12, 24, 48, 72}),
+			},
 		},
 	}
-}
-
-func expandInfraAlertCondition(d *schema.ResourceData) (*alerts.InfrastructureCondition, error) {
-	condition := alerts.InfrastructureCondition{
-		Name:       d.Get("name").(string),
-		Enabled:    d.Get("enabled").(bool),
-		PolicyID:   d.Get("policy_id").(int),
-		Event:      d.Get("event").(string),
-		Comparison: d.Get("comparison").(string),
-		Select:     d.Get("select").(string),
-		Type:       d.Get("type").(string),
-		Critical:   expandAlertThreshold(d.Get("critical")),
-	}
-
-	if attr, ok := d.GetOk("runbook_url"); ok {
-		condition.RunbookURL = attr.(string)
-	}
-	if attr, ok := d.GetOk("warning"); ok {
-		condition.Warning = expandAlertThreshold(attr)
-	}
-
-	if attr, ok := d.GetOk("where"); ok {
-		condition.Where = attr.(string)
-	}
-
-	if attr, ok := d.GetOk("process_where"); ok {
-		condition.ProcessWhere = attr.(string)
-	}
-
-	if attr, ok := d.GetOk("integration_provider"); ok {
-		condition.IntegrationProvider = attr.(string)
-	}
-
-	err := validateAttributesForType(&condition)
-
-	if err != nil {
-		return nil, err
-	}
-
-	return &condition, nil
-}
-
-func validateAttributesForType(c *alerts.InfrastructureCondition) error {
-	switch c.Type {
-	case "infra_process_running":
-		if c.Event != "" {
-			return fmt.Errorf("event is not supported by condition type %s", c.Type)
-		}
-		if c.IntegrationProvider != "" {
-			return fmt.Errorf("integration_provider is not supported by condition type %s", c.Type)
-		}
-		if c.Select != "" {
-			return fmt.Errorf("select is not supported by condition type %s", c.Type)
-		}
-		if c.Critical.Function != "" {
-			return fmt.Errorf("time_function is not supported by condition type %s", c.Type)
-		}
-	case "infra_metric":
-		if c.ProcessWhere != "" {
-			return fmt.Errorf("process_where is not supported by condition type %s", c.Type)
-		}
-	case "infra_host_not_reporting":
-		if c.Event != "" {
-			return fmt.Errorf("event is not supported by condition type %s", c.Type)
-		}
-		if c.IntegrationProvider != "" {
-			return fmt.Errorf("integration_provider is not supported by condition type %s", c.Type)
-		}
-		if c.Select != "" {
-			return fmt.Errorf("select is not supported by condition type %s", c.Type)
-		}
-		if c.ProcessWhere != "" {
-			return fmt.Errorf("process_where is not supported by condition type %s", c.Type)
-		}
-		if c.Comparison != "" {
-			return fmt.Errorf("comparison is not supported by condition type %s", c.Type)
-		}
-		if c.Critical.Function != "" {
-			return fmt.Errorf("time_function is not supported by condition type %s", c.Type)
-		}
-		if c.Critical.Value != 0 {
-			return fmt.Errorf("value is not supported by condition type %s", c.Type)
-		}
-	}
-
-	return nil
-}
-
-func readInfraAlertConditionStruct(condition *alerts.InfrastructureCondition, d *schema.ResourceData) error {
-	ids, err := parseIDs(d.Id(), 2)
-	if err != nil {
-		return err
-	}
-
-	policyID := ids[0]
-
-	d.Set("policy_id", policyID)
-	d.Set("name", condition.Name)
-	d.Set("runbook_url", condition.RunbookURL)
-	d.Set("enabled", condition.Enabled)
-	d.Set("comparison", condition.Comparison)
-	d.Set("event", condition.Event)
-	d.Set("select", condition.Select)
-	d.Set("type", condition.Type)
-	d.Set("created_at", condition.CreatedAt)
-	d.Set("updated_at", condition.UpdatedAt)
-
-	if condition.Where != "" {
-		d.Set("where", condition.Where)
-	}
-
-	if condition.ProcessWhere != "" {
-		d.Set("process_where", condition.ProcessWhere)
-	}
-
-	if condition.IntegrationProvider != "" {
-		d.Set("integration_provider", condition.IntegrationProvider)
-	}
-
-	if err := d.Set("critical", flattenAlertThreshold(condition.Critical)); err != nil {
-		return err
-	}
-
-	if condition.Warning != nil {
-		if err := d.Set("warning", flattenAlertThreshold(condition.Warning)); err != nil {
-			return err
-		}
-	}
-
-	return nil
 }
 
 func resourceNewRelicInfraAlertConditionCreate(d *schema.ResourceData, meta interface{}) error {
@@ -323,7 +197,7 @@ func resourceNewRelicInfraAlertConditionRead(d *schema.ResourceData, meta interf
 		return err
 	}
 
-	return readInfraAlertConditionStruct(condition, d)
+	return flattenInfraAlertCondition(condition, d)
 }
 
 func resourceNewRelicInfraAlertConditionUpdate(d *schema.ResourceData, meta interface{}) error {
@@ -372,32 +246,4 @@ func resourceNewRelicInfraAlertConditionDelete(d *schema.ResourceData, meta inte
 	}
 
 	return nil
-}
-
-func expandAlertThreshold(v interface{}) *alerts.InfrastructureConditionThreshold {
-	rah := v.([]interface{})[0].(map[string]interface{})
-
-	alertInfraThreshold := &alerts.InfrastructureConditionThreshold{
-		Duration: rah["duration"].(int),
-	}
-
-	if val, ok := rah["value"]; ok {
-		alertInfraThreshold.Value = val.(float64)
-	}
-
-	if val, ok := rah["time_function"]; ok {
-		alertInfraThreshold.Function = val.(string)
-	}
-
-	return alertInfraThreshold
-}
-
-func flattenAlertThreshold(v *alerts.InfrastructureConditionThreshold) []interface{} {
-	alertInfraThreshold := map[string]interface{}{
-		"duration":      v.Duration,
-		"value":         v.Value,
-		"time_function": v.Function,
-	}
-
-	return []interface{}{alertInfraThreshold}
 }
